pyproject.toml
[project]
name = "rag-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.5.3",
    "ipykernel>=7.2.0",
    "langchain-openrouter>=0.2.1",
    "langchain[google-genai]>=1.2.15",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
